backend/ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-cerebras==0.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.2.1  # Recommended for better configuration handling
psycopg2-binary==2.9.9    # Recommended over regular psycopg2 for Docker builds
httpx==0.27.0             # Recommended for stable async HTTP requests
seaborn==0.13.2
matplotlib==3.10.6